security+feature: shield recovery system, TX checked instructions, hardening fixes
Shield Recovery System:
- Manual shield recovery (24h window instead of auto-consume)
- POST /api/v1/shields/recover endpoint
- Recovery banner in HomeTab + PerksListScreen
- Scheduler sets recovery window on missed burn days
- Migration 0005: streak_recovery_deadline, streak_recovery_gap_days columns

Transaction Security (Android):
- Switched all SPL instructions to Checked variants (TransferChecked/BurnChecked)
- Wallet now displays token type, amount, destination in approval screen
- Eliminates 'transaction can be simulated' warning on Seeker wallet

Backend Security Hardening:
- Shop: source account validation on SPL transfers, account index bounds check
- Deposit: buffer bounds check + account index validation
- Rate limits: fail-closed (503) instead of fail-open when Redis is down
- Price tolerance: 15% -> 5%
- Leaderboard: in-process fallback cache against stampede when Redis is down
- recheckDailySweep: pass walletAddress instead of userId
- Schema: xp column type aligned to BIGINT (matches migration)
- Migration 0002: price_lamports type fixed to NUMERIC(20,0)
- Env fallback prices updated (SOL=, SKR=.017)

Credential Cleanup:
- Removed hardcoded Railway DB password from run-mint-migration.ts

+++ backend/migrations/0002_xp_challenges_shop.sql
@@ -5,6 +5,7 @@
 -- XP columns on users table
 -- ──────────────────────────────────────────
 ALTER TABLE users ADD COLUMN IF NOT EXISTS xp BIGINT NOT NULL DEFAULT 0;
+-- NOTE: Drizzle schema uses bigint({ mode: 'number' }) for this column.
 ALTER TABLE users ADD COLUMN IF NOT EXISTS level INTEGER NOT NULL DEFAULT 1;
 ALTER TABLE users ADD COLUMN IF NOT EXISTS streak_shields INTEGER NOT NULL DEFAULT 0;
 ALTER TABLE users ADD COLUMN IF NOT EXISTS profile_title VARCHAR(40);
@@ -68,7 +69,7 @@ CREATE TABLE IF NOT EXISTS shield_purchases (
   wallet_address  VARCHAR(44) NOT NULL,
   tx_signature    VARCHAR(88) NOT NULL UNIQUE,
   shield_count    INTEGER NOT NULL,
-  price_lamports  BIGINT NOT NULL,
+  price_lamports  NUMERIC(20,0) NOT NULL,
   price_usd       NUMERIC(10,2) NOT NULL,
   status          VARCHAR(20) NOT NULL DEFAULT 'PENDING',  -- PENDING, VERIFIED, FAILED
   verified_at     TIMESTAMPTZ,
